Prepare the next release.
