+++ .github/workflows/test.yml
@@ -11,8 +11,6 @@ on:
 
 jobs:
   test:
-    name: Test
-
     strategy:
       matrix:
         os: [ubuntu, windows, macos]
@@ -48,3 +46,4 @@ jobs:
         uses: codecov/codecov-action@v4
         with:
           files: "./coverage.xml"
+          token: ${{ secrets.CODECOV_TOKEN }}
