+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.3.3
+  rev: v0.4.10
   hooks:
   - id: ruff
     name: Run the Ruff linter
@@ -11,7 +11,7 @@ repos:
     types_or: [python, pyi]
 
 - repo: https://github.com/python-poetry/poetry
-  rev: 1.8.2
+  rev: 1.8.3
   hooks:
   - id: poetry-lock
     stages: [push]
