--- Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "extension"
-version = "0.1.0"
-authors = ["nekitdev", "RealistikDash"]
-edition = "2021"
-description = "Simple, reusable and optimized XOR ciphers in Python (extension)."
-readme = "README.md"
-documentation = "https://docs.xor-cipher.org/"
-homepage = "https://xor-cipher.org/"
-repository = "https://github.com/xor-cipher/xor-cipher"
-license = "MIT"
-keywords = ["xor", "cipher", "extension"]
-publish = false
-
-[lib]
-name = "extension"
-crate-type = ["cdylib"]
-
-[dependencies.pyo3]
-version = ">= 0.20.3"
-features = ["extension-module"]
