+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "xor-cipher"
-version = "4.0.0"
+version = "5.0.0"
 description = "Simple, reusable and optimized XOR ciphers in Python."
 authors = ["nekitdev", "RealistikDash"]
 license = "MIT"
@@ -24,63 +24,43 @@ classifiers = [
 [tool.poetry.urls]
 Issues = "https://github.com/xor-cipher/xor-cipher/issues"
 
-[[tool.poetry.packages]]
-include = "xor_cipher"
-
-[[tool.poetry.include]]
-path = "xor_cipher/*.so"
-format = "wheel"
-
-[[tool.poetry.include]]
-path = "xor_cipher/*.pyd"
-format = "wheel"
-
-[[tool.poetry.include]]
-path = "xor_cipher"
-format = "wheel"
-
-[tool.poetry.build]
-generate-setup-file = false
-script = "build.py"
-
 [tool.poetry.dependencies]
 python = ">= 3.8"
-annotated-types = ">= 0.6.0"
-typing-extensions = ">= 4.10.0"
+
+xor-cipher-core = ">= 1.1.0"
 
 [tool.poetry.group.format.dependencies]
-ruff = "0.3.3"
+ruff = "0.4.10"
 
 [tool.poetry.group.check.dependencies]
-mypy = "1.9.0"
+mypy = "1.10.1"
 
 [tool.poetry.group.check.dependencies.pre-commit]
-version = "3.6.2"
+version = "3.7.1"
 python = ">= 3.9"
 
 [tool.poetry.group.test.dependencies]
-coverage = "7.4.4"
-pytest = "8.1.1"
-pytest-cov = "4.1.0"
-hypothesis = "6.99.6"
-funcs = "0.10.0"
+coverage = "7.5.4"
+pytest = "8.2.2"
+pytest-cov = "5.0.0"
+hypothesis = "6.104.1"
 
 [tool.poetry.group.docs]
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-mkdocs = "1.5.3"
-mkdocs-material = "9.5.13"
+mkdocs = "1.6.0"
+mkdocs-material = "9.5.27"
 
 [tool.poetry.group.docs.dependencies.mkdocstrings]
-version = "0.24.0"
+version = "0.25.1"
 extras = ["python"]
 
 [tool.poetry.group.release]
 optional = true
 
 [tool.poetry.group.release.dependencies]
-changelogging = "1.4.2"
+changelogging = "2.3.0"
 
 [tool.ruff]
 line-length = 100
@@ -112,26 +92,15 @@ directory = "coverage"
 [tool.mypy]
 strict = true
 
-exclude = ["build.py"]
-
-[tool.changelogging]
+[tool.changelogging.context]
 name = "xor-cipher"
-version = "4.0.0"
+version = "5.0.0"
 url = "https://github.com/xor-cipher/xor-cipher"
-directory = "changes"
-output = "CHANGELOG.md"
 
-start_string = "<!-- changelogging: start -->"
-
-title_format = "{version} ({date})"
-issue_format = "[#{issue}]({url}/pull/{issue})"
-
-bullet = "-"
-wrap = true
-wrap_size = 100
-
-display = ["feature", "change", "fix", "security", "deprecation", "removal", "internal"]
+[tool.changelogging.formats]
+title = "[{{version}}]({{url}}/tree/v{{version}}) ({{date}})"
+fragment = "{{content}} ([#{{id}}]({{url}}/pull/{{id}}))"
 
 [build-system]
-requires = ["poetry-core >= 1.9.0", "maturin >= 1.5.0", "entrypoint >= 2.0.3"]
+requires = ["poetry-core >= 1.9.0"]
 build-backend = "poetry.core.masonry.api"
