--- xor_cipher/extension.pyi
@@ -1,6 +1,0 @@
-from xor_cipher.typing import byte
-
-def xor(data: bytes, key: byte) -> bytes: ...
-def cyclic_xor(data: bytes, key: bytes) -> bytes: ...
-def xor_in_place(data: bytearray, key: byte) -> None: ...
-def cyclic_xor_in_place(data: bytearray, key: bytes) -> None: ...
